requirements.txt
google-generativeai
google-genai
python-dotenv
python-pptx